move offline screens to offline module
