+++ core/data/src/main/java/com/mifos/core/data/repository_imp/SyncClientPayloadsRepositoryImp.kt
@@ -1,5 +1,6 @@
-package com.mifos.mifosxdroid.offline.syncclientpayloads
+package com.mifos.core.data.repository_imp
 
+import com.mifos.core.data.repository.SyncClientPayloadsRepository
 import com.mifos.core.network.datamanager.DataManagerClient
 import com.mifos.core.objects.client.Client
 import com.mifos.core.objects.client.ClientPayload
