+++ feature/offline/src/main/java/com/mifos/feature/offline/sync_savings_account_transaction/SyncSavingsAccountTransactionScreen.kt
@@ -1,4 +1,4 @@
-package com.mifos.mifosxdroid.offline.syncloanrepaymenttransacition
+package com.mifos.feature.offline.sync_savings_account_transaction
 
 import android.content.Context
 import android.widget.Toast
@@ -6,10 +6,8 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
@@ -17,8 +15,8 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AssignmentTurnedIn
 import androidx.compose.material.icons.filled.Error
+import androidx.compose.material.icons.filled.Sync
 import androidx.compose.material.pullrefresh.PullRefreshIndicator
 import androidx.compose.material.pullrefresh.pullRefresh
 import androidx.compose.material.pullrefresh.rememberPullRefreshState
@@ -33,84 +31,87 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.mifos.core.common.utils.Network
 import com.mifos.core.designsystem.component.MifosCircularProgress
 import com.mifos.core.designsystem.component.MifosScaffold
 import com.mifos.core.designsystem.icon.MifosIcons
 import com.mifos.core.objects.PaymentTypeOption
-import com.mifos.core.objects.accounts.loan.LoanRepaymentRequest
-import com.mifos.mifosxdroid.R
-import com.mifos.utils.Network
-import com.mifos.utils.PrefManager.userStatus
-import com.mifos.utils.Utils.getPaymentTypeName
+import com.mifos.core.objects.accounts.savings.SavingsAccountTransactionRequest
+import com.mifos.core.ui.components.MifosEmptyUi
+import com.mifos.feature.offline.R
+import com.mifos.feature.savings.sync_account_transaction.SyncSavingsAccountTransactionUiState
+import com.mifos.feature.savings.sync_account_transaction.SyncSavingsAccountTransactionViewModel
 
 @Composable
-fun SyncLoanRepaymentTransactionScreenRoute(
-    viewModel: SyncLoanRepaymentTransactionViewModel = hiltViewModel(),
+fun SyncSavingsAccountTransactionScreenRoute(
+    viewModel: SyncSavingsAccountTransactionViewModel = hiltViewModel(),
     onBackPressed: () -> Unit,
 ) {
-    val uiState by viewModel.syncLoanRepaymentTransactionUiState.collectAsStateWithLifecycle()
+    val uiState by viewModel.syncSavingsAccountTransactionUiState.collectAsStateWithLifecycle()
     val refreshState by viewModel.isRefreshing.collectAsStateWithLifecycle()
 
     LaunchedEffect(key1 = Unit) {
-        viewModel.loadDatabaseLoanRepaymentTransactions()
-        viewModel.loanPaymentTypeOption()
+        viewModel.loadDatabaseSavingsAccountTransactions()
+        viewModel.loadPaymentTypeOption()
     }
 
-    SyncLoanRepaymentTransactionScreen(
+    SyncSavingsAccountTransactionScreen(
         uiState = uiState,
         onBackPressed = onBackPressed,
         refreshState = refreshState,
         onRefresh = {
             viewModel.refreshTransactions()
         },
-        syncLoanRepaymentTransactions = {
-            viewModel.syncGroupPayload()
-        }
+        syncSavingsAccountTransactions = {
+            viewModel.syncSavingsAccountTransactions()
+        },
+        getUserStatus = {viewModel.getUserStatus()}
     )
 }
 
+
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
-fun SyncLoanRepaymentTransactionScreen(
-    uiState: SyncLoanRepaymentTransactionUiState,
+fun SyncSavingsAccountTransactionScreen(
+    uiState: SyncSavingsAccountTransactionUiState,
     onBackPressed: () -> Unit,
     refreshState: Boolean,
     onRefresh: () -> Unit,
-    syncLoanRepaymentTransactions: () -> Unit,
+    syncSavingsAccountTransactions: () -> Unit,
+    getUserStatus: () -> Boolean
 ) {
-    val snackbarHostState = remember { SnackbarHostState() }
+    val snackbarHostState by remember { mutableStateOf(SnackbarHostState()) }
     val context = LocalContext.current
-    val pullRefreshState =
-        rememberPullRefreshState(refreshing = refreshState, onRefresh = onRefresh)
+    val pullRefreshState = rememberPullRefreshState(
+        refreshing = refreshState,
+        onRefresh = onRefresh
+    )
 
     MifosScaffold(
         icon = MifosIcons.arrowBack,
-        title = stringResource(id = R.string.sync_loanrepayment),
+        title = stringResource(id = R.string.feature_offline_sync_savingsAccountTransactions),
         onBackPressed = onBackPressed,
         actions = {
             IconButton(onClick = {
-                when (userStatus) {
-                    false -> checkNetworkConnectionAndSync(context, syncLoanRepaymentTransactions)
-                    true -> TODO("Implement OfflineModeDialog()")
+                when (getUserStatus()) {
+                    false -> checkNetworkConnectionAndSync(context, syncSavingsAccountTransactions)
+                    true -> TODO()//Implement OfflineModeDialog()
                 }
             }) {
-                Icon(
-                    MifosIcons.sync,
-                    contentDescription = stringResource(id = R.string.sync_loanrepayment)
-                )
+                Icon(Icons.Default.Sync, contentDescription = "Sync")
             }
         },
         snackbarHostState = snackbarHostState
@@ -118,23 +119,30 @@ fun SyncLoanRepaymentTransactionScreen(
         Column(modifier = Modifier.padding(paddingValues)) {
             Box(modifier = Modifier.pullRefresh(pullRefreshState)) {
                 when (uiState) {
-                    is SyncLoanRepaymentTransactionUiState.ShowProgressbar -> {
+                    is SyncSavingsAccountTransactionUiState.Loading -> {
                         MifosCircularProgress()
                     }
 
-                    is SyncLoanRepaymentTransactionUiState.ShowError -> {
-                        ErrorStateScreen(uiState.message.toString(), onRefresh)
-                    }
-
-                    is SyncLoanRepaymentTransactionUiState.ShowLoanRepaymentTransactions -> {
-                        LoanRepaymentTransactionsList(
-                            uiState.loanRepaymentRequests,
-                            uiState.paymentTypeOptions
+                    is SyncSavingsAccountTransactionUiState.ShowEmptySavingsAccountTransactions -> {
+                        MifosEmptyUi(
+                            text = stringResource(id = R.string.feature_offline_nothing_to_sync),
+                            icon = MifosIcons.sync
                         )
                     }
 
-                    is SyncLoanRepaymentTransactionUiState.ShowEmptyLoanRepayments -> {
-                        EmptyLoanRepaymentsScreen(uiState.message)
+                    is SyncSavingsAccountTransactionUiState.ShowError -> {
+                        val message = uiState.message
+                        ErrorStateScreen(message.toString(), onRefresh)
+                    }
+
+                    is SyncSavingsAccountTransactionUiState.ShowSavingsAccountTransactions -> {
+                        val transactions = uiState.savingsList
+                        val paymentTypeOptions = uiState.paymentTypeOptions
+                        LazyColumn {
+                            items(transactions) { transaction ->
+                                SavingsAccountTransactionItem(transaction, paymentTypeOptions)
+                            }
+                        }
                     }
                 }
                 PullRefreshIndicator(
@@ -148,20 +156,8 @@ fun SyncLoanRepaymentTransactionScreen(
 }
 
 @Composable
-fun LoanRepaymentTransactionsList(
-    loanRepaymentRequests: List<LoanRepaymentRequest>,
-    paymentTypeOptions: List<PaymentTypeOption>
-) {
-    LazyColumn {
-        items(loanRepaymentRequests) { request ->
-            LoanRepaymentTransactionItem(request, paymentTypeOptions)
-        }
-    }
-}
-
-@Composable
-fun LoanRepaymentTransactionItem(
-    request: LoanRepaymentRequest,
+fun SavingsAccountTransactionItem(
+    transaction: SavingsAccountTransactionRequest,
     paymentTypeOptions: List<PaymentTypeOption>
 ) {
     Card(
@@ -176,27 +172,33 @@ fun LoanRepaymentTransactionItem(
                 .fillMaxWidth()
                 .padding(8.dp)
         ) {
-            TransactionRow(stringResource(R.string.loan_id), request.loanId.toString())
-            TransactionRow(stringResource(R.string.account_number), request.accountNumber ?: "")
             TransactionRow(
-                stringResource(R.string.payment_type),
-                request.paymentTypeId?.let { getPaymentTypeName(it.toInt(), paymentTypeOptions) }
-                    ?: ""
+                label = stringResource(R.string.feature_offline_savings_account_id),
+                value = transaction.savingAccountId.toString()
             )
             TransactionRow(
-                stringResource(R.string.transaction_amount),
-                request.transactionAmount ?: ""
+                label = stringResource(R.string.feature_offline_payment_type),
+                value = transaction.paymentTypeId?.toInt()?.let {
+                    getPaymentTypeName(it, paymentTypeOptions)
+                } ?: ""
             )
             TransactionRow(
-                stringResource(R.string.loan_transaction_date),
-                request.transactionDate ?: ""
+                label = stringResource(R.string.feature_offline_transaction_type),
+                value = transaction.transactionType ?: ""
+            )
+            TransactionRow(
+                label = stringResource(R.string.feature_offline_transaction_amount),
+                value = transaction.transactionAmount ?: ""
+            )
+            TransactionRow(
+                label = stringResource(R.string.feature_offline_transaction_date),
+                value = transaction.transactionDate ?: ""
             )
 
-            if (request.errorMessage != null) {
+            if (transaction.errorMessage != null) {
                 Text(
-                    text = request.errorMessage!!,
+                    text = transaction.errorMessage!!,
                     style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.error,
                     modifier = Modifier.padding(top = 4.dp)
                 )
             }
@@ -220,6 +222,7 @@ fun TransactionRow(label: String, value: String) {
         Text(
             text = value,
             style = MaterialTheme.typography.bodyMedium,
+            fontWeight = FontWeight.Bold,
             modifier = Modifier.weight(1f)
         )
     }
@@ -237,61 +240,46 @@ fun ErrorStateScreen(message: String, onRefresh: () -> Unit) {
             contentDescription = null,
             modifier = Modifier.size(48.dp)
         )
-        Text(text = message, modifier = Modifier.padding(vertical = 6.dp))
+        Text(text = message, modifier = Modifier.padding(vertical = 8.dp))
         Button(onClick = onRefresh) {
-            Text(stringResource(id = R.string.click_to_refresh))
+            Text(stringResource(id = R.string.feature_offline_retry))
         }
     }
 }
 
-@Composable
-fun EmptyLoanRepaymentsScreen(message: String) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Icon(
-            imageVector = Icons.Default.AssignmentTurnedIn,
-            contentDescription = null,
-            modifier = Modifier.size(72.dp),
-            tint = MaterialTheme.colorScheme.primary
-        )
-        Spacer(modifier = Modifier.height(16.dp))
-        Text(
-            text = message,
-            style = MaterialTheme.typography.bodyLarge,
-            textAlign = TextAlign.Center,
-            color = MaterialTheme.colorScheme.onSurface
-        )
-    }
+fun getPaymentTypeName(
+    paymentId: Int,
+    paymentTypeOptions: List<PaymentTypeOption>?
+): String? {
+    return paymentTypeOptions
+        ?.firstOrNull { it.id == paymentId }
+        ?.name
 }
 
 fun checkNetworkConnectionAndSync(
     context: Context,
-    syncLoanRepaymentTransactions: () -> Unit
+    syncSavingsAccountTransactions: () -> Unit
 ) {
     if (Network.isOnline(context)) {
-        syncLoanRepaymentTransactions()
+        syncSavingsAccountTransactions()
     } else {
         Toast.makeText(
             context,
-            context.getString(R.string.error_not_connected_internet),
+            context.resources.getString(R.string.feature_offline_error_not_connected_internet),
             Toast.LENGTH_SHORT
         ).show()
     }
 }
 
-class SyncLoanRepaymentTransactionUiStateProvider :
-    PreviewParameterProvider<SyncLoanRepaymentTransactionUiState> {
+
+class SyncSavingsAccountTransactionUiStateProvider :
+    PreviewParameterProvider<SyncSavingsAccountTransactionUiState> {
     override val values = sequenceOf(
-        SyncLoanRepaymentTransactionUiState.ShowProgressbar,
-        SyncLoanRepaymentTransactionUiState.ShowError(R.string.failed_to_load_loanrepayment),
-        SyncLoanRepaymentTransactionUiState.ShowEmptyLoanRepayments("No loan repayments to sync"),
-        SyncLoanRepaymentTransactionUiState.ShowLoanRepaymentTransactions(
-            sampleLoanRepaymentRequests,
+        SyncSavingsAccountTransactionUiState.Loading,
+        SyncSavingsAccountTransactionUiState.ShowError(R.string.feature_offline_failed_to_load_savingaccounttransaction),
+        SyncSavingsAccountTransactionUiState.ShowEmptySavingsAccountTransactions(R.string.feature_offline_no_transaction_to_sync),
+        SyncSavingsAccountTransactionUiState.ShowSavingsAccountTransactions(
+            sampleSavingsAccountTransactions.toMutableList(),
             samplePaymentTypeOptions
         )
     )
@@ -299,27 +287,32 @@ class SyncLoanRepaymentTransactionUiStateProvider :
 
 @Preview(showBackground = true)
 @Composable
-private fun SyncLoanRepaymentTransactionScreenPreview(
-    @PreviewParameter(SyncLoanRepaymentTransactionUiStateProvider::class) uiState: SyncLoanRepaymentTransactionUiState
+private fun SyncSavingsAccountTransactionScreenPreview(
+    @PreviewParameter(SyncSavingsAccountTransactionUiStateProvider::class) state: SyncSavingsAccountTransactionUiState
 ) {
-    SyncLoanRepaymentTransactionScreen(
-        uiState = uiState,
+    SyncSavingsAccountTransactionScreen(
+        uiState = state,
         onBackPressed = {},
         refreshState = false,
         onRefresh = {},
-        syncLoanRepaymentTransactions = {}
+        syncSavingsAccountTransactions = {},
+        getUserStatus = { true }
     )
 }
 
 // Sample data for previews
-val sampleLoanRepaymentRequests = List(5) { index ->
-    LoanRepaymentRequest(
-        loanId = index,
-        accountNumber = "LOAN-$index",
-        paymentTypeId = index.toString(),
-        transactionAmount = "${1000 + index * 100}",
+val sampleSavingsAccountTransactions = List(5) { index ->
+    SavingsAccountTransactionRequest(
+        savingAccountId = index,
         transactionDate = "2023-07-${15 + index}",
-        errorMessage = if (index % 2 == 0) null else "Error in transaction"
+        transactionAmount = "${100 + index * 10}",
+        paymentTypeId = index.toLong().toString(),
+        transactionType = if (index % 2 == 0) "deposit" else "withdrawal",
+        accountNumber = "ACC-$index",
+        checkNumber = "CHK-$index",
+        routingCode = "RTG-$index",
+        receiptNumber = "RCP-$index",
+        bankNumber = "BNK-$index"
     )
 }
 
@@ -333,12 +326,12 @@ val samplePaymentTypeOptions = List(3) { index ->
     )
 }
 
-// Individual preview for LoanRepaymentTransactionItem
+// Individual preview for SavingsAccountTransactionItem
 @Preview(showBackground = true)
 @Composable
-fun LoanRepaymentTransactionItemPreview() {
-    LoanRepaymentTransactionItem(
-        request = sampleLoanRepaymentRequests[0],
+fun SavingsAccountTransactionItemPreview() {
+    SavingsAccountTransactionItem(
+        transaction = sampleSavingsAccountTransactions[0],
         paymentTypeOptions = samplePaymentTypeOptions
     )
 }
