+++ feature/offline/src/main/res/drawable/feature_offline_ic_assignment_turned_in_black_24dp.xml
