+++ feature/offline/src/main/res/values/strings.xml
@@ -3,9 +3,57 @@
     <string name="feature_offline_nothing_to_sync">Nothing To Sync :)</string>
     <string name="feature_offline_offline_Sync">Offline Sync *</string>
     <string name="feature_offline_sync_loanRepayments">LoanRepayments</string>
-    <string name="feature_offline_sync_groups">Groups</string>
     <string name="feature_offline_sync_clients">Clients</string>
     <string name="feature_offline_sync_centers">Centers</string>
     <string name="feature_offline_sync_savingsAccountTransactions">Sync Savings Account Transactions</string>
+
+    <string name="feature_offline_sync_clients_payloads">Sync Client Payloads</string>
+    <string name="feature_offline_first_name">First name</string>
+    <string name="feature_offline_middle_name">Middle name</string>
+    <string name="feature_offline_last_name">Last name</string>
+    <string name="feature_offline_mobile_no">Mobile Number</string>
+    <string name="feature_offline_gender">Gender</string>
+    <string name="feature_offline_dob">Date of Birth</string>
+    <string name="feature_offline_activation_date">Activation Date</string>
+    <string name="feature_offline_active">Active</string>
+    <string name="feature_offline_error_not_connected_internet">Not connected to Network</string>
+
+    <string name="feature_offline_sync_loanrepayment">Sync Loan Repayment Transaction</string>
+    <string name="feature_offline_loan_id">Loan ID</string>
+    <string name="feature_offline_account_number">Account Number</string>
+    <string name="feature_offline_transaction_amount">Transaction amount*</string>
+    <string name="feature_offline_loan_transaction_date">Loan Transaction Date</string>
+    <string name="feature_offline_click_to_refresh">\n Click to Refresh</string>
+    <string name="feature_offline_failed_to_load_loanrepayment">Failed to load LoanRepayment</string>
+
+    <string name="feature_offline_sync_savingsaccounttransactions">Sync Savings Account Transactions</string>
+    <string name="feature_offline_savings_account_id">Savings Account Id</string>
+    <string name="feature_offline_payment_type">Payment Type</string>
+    <string name="feature_offline_transaction_type">Transaction Type</string>
+    <string name="feature_offline_transaction_date">Transaction Date</string>
+    <string name="feature_offline_retry">Retry</string>
+    <string name="feature_offline_failed_to_load_savingaccounttransaction">Failed to loadSavingsAccountTransaction</string>
+    <string name="feature_offline_error_fix_before_sync">Please Fix the Error before sync</string>
+    <string name="feature_offline_failed_to_load_paymentoptions">Failed to Load PaymentTypeOptions</string>
+    <string name="feature_offline_no_transaction_to_sync">There is No Transaction to Sync</string>
+    <string name="feature_offline_failed_to_update_list">Failed to Update List</string>
+    <string name="feature_offline_failed_to_update_savingsaccount">Failed to update Savings Accounts</string>
+
+    <string name="feature_offline_sync_groups">Groups</string>
+    <string name="feature_offline_sync">Sync</string>
+    <string name="feature_offline_name">Name</string>
+    <string name="feature_offline_external_id">External ID</string>
+    <string name="feature_offline_office_id">Office Id</string>
+    <string name="feature_offline_submit_date">Submit Date</string>
+
+    <string name="feature_offline_all_groups_synced">All Groups have been Synced</string>
+    <string name="feature_offline_no_group_payload_to_sync">There is No Group Payload to Sync</string>
+    <string name="feature_offline_error_failed_to_update_list">Failed to Update List</string>
+    <string name="feature_offline_error_group_sync_failed">Group sync failed</string>
+    <string name="feature_offline_error_failed_to_load_groupPayload">Failed to load GroupPayload</string>
+
+    <string name="feature_offline_sync_centers_payloads">Sync Centers Payloads</string>
+    <string name="feature_offline_no_center_payload_to_sync">There is No Center Payload to Sync</string>
+    <string name="feature_offline_no_loanrepayment_to_sync">No Loan Repayment to sync</string>
 
 </resources>
