+++ mifosng-android/src/main/java/com/mifos/mifosxdroid/components/Navigation.kt
@@ -38,8 +38,8 @@ import com.mifos.feature.loan.navigation.navigateToLoanAccountScreen
 import com.mifos.feature.loan.navigation.navigateToLoanAccountSummaryScreen
 import com.mifos.feature.note.navigation.navigateToNoteScreen
 import com.mifos.feature.note.navigation.noteScreen
-import com.mifos.feature.offline.dashboard.navigation.navigateToOfflineDashboardScreen
-import com.mifos.feature.offline.dashboard.navigation.offlineNavGraph
+import com.mifos.feature.offline.navigation.navigateToOfflineDashboardScreen
+import com.mifos.feature.offline.navigation.offlineNavGraph
 import com.mifos.feature.path_tracking.navigation.pathTrackingNavGraph
 import com.mifos.feature.report.navigation.reportNavGraph
 import com.mifos.feature.savings.navigation.addSavingsAccountScreen
