+++ mifosng-android/src/main/java/com/mifos/mifosxdroid/offline/syncgrouppayloads/SyncGroupPayloadsFragment.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.navigation.findNavController
-import com.mifos.feature.groups.sync_group_payloads.SyncGroupPayloadsScreenRoute
+import com.mifos.feature.offline.sync_group_payloads.SyncGroupPayloadsScreenRoute
 import com.mifos.mifosxdroid.core.MifosBaseFragment
 import dagger.hilt.android.AndroidEntryPoint
